improv: moved Network.url: String to a better place
