+++ BDKSwiftExampleWallet/Utilities/Constants.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import SwiftUI
+import BitcoinDevKit
 
 struct Constants {
     struct Config {
@@ -73,6 +74,23 @@ struct Constants {
             case .testnet4:
                 return Color.cyan
             }
+        }
+    }
+}
+
+extension Network {
+    var url: String {
+        switch self {
+        case .bitcoin:
+            Constants.Config.EsploraServerURLNetwork.Bitcoin.allValues.first ?? ""
+        case .testnet:
+            Constants.Config.EsploraServerURLNetwork.Testnet.allValues.first ?? ""
+        case .signet:
+            Constants.Config.EsploraServerURLNetwork.Signet.allValues.first ?? ""
+        case .regtest:
+            Constants.Config.EsploraServerURLNetwork.Regtest.allValues.first ?? ""
+        case .testnet4:
+            Constants.Config.EsploraServerURLNetwork.Testnet4.allValues.first ?? ""
         }
     }
 }
